+++ ChangeLog.md
@@ -4,7 +4,8 @@ T-Regx Changelog
 Incoming
 --------
 
-* Soon
+* Bug fixes
+    * Fixed a bug when pattern `[]]` with prepared pattern was parsed incorrectly.
 
 Added in 0.19.0
 ---------------